Feature/73 remote task

+++ tschiraplus/Services/TaskServices/TaskService.cs
@@ -5,6 +5,7 @@ using Core.Models;
 using Services.DTOs;
 using Services.Repositories;
 using TaskStatus = Core.Enums.TaskStatus;
+using TaskPriority = Core.Enums.TaskPriority;
 
 namespace Services.TaskServices;
 
@@ -64,8 +65,8 @@ public class TaskService : ITaskService
             "status" => _taskSortingManager.FilterByPredicate(tasks,
                 t => t.Status == value),
             
-            "creationdate" => _taskSortingManager.FilterByPredicate(tasks,
-                t => t.CreationDate.Date == DateTime.Parse(value).Date),
+            "startdate" => _taskSortingManager.FilterByPredicate(tasks,
+                t => t.StartDate == DateTime.Parse(value).Date),
             
             "priority" => _taskSortingManager.FilterByPredicate(tasks,
                 t => t.DueDate == DateTime.Parse(value).Date),
@@ -83,6 +84,7 @@ public class TaskService : ITaskService
     public void CreateTask(TaskDto task)
     {
         _taskRepository.AddTask(ConvertTaskDtoToTaskModel(task));
+        _taskRepository.PostTaskAsync(ConvertTaskDtoToTaskModel(task));
     }
     
     /// <summary>
@@ -93,7 +95,7 @@ public class TaskService : ITaskService
     /// <param name="status"></param>
     /// <param name="creationDate"></param>
     /// <returns>The newly created TaskDto</returns>
-    public TaskDto CreateTaskDto(string title, string description, string status, DateTime creationDate)
+    public TaskDto CreateTaskDto(string title, string description, string status, string priority, DateTime creationDate)
     {
         var dto = new TaskDto
         {
@@ -101,7 +103,8 @@ public class TaskService : ITaskService
             Title = title,
             Description = description,
             Status = status,
-            CreationDate = creationDate
+            Priority = priority,
+            StartDate = null
         };
         return dto;
     }    
@@ -121,6 +124,7 @@ public class TaskService : ITaskService
             Title = taskDto.Title,
             Description = taskDto.Description,
             Status = Enum.TryParse(taskDto.Status, out TaskStatus status) ? status : TaskStatus.Backlog,
+            Priority = Enum.TryParse(taskDto.Priority, out TaskPriority priority) ? priority : TaskPriority.High,
             CreationDate = DateTime.Now,
             LastUpdated = DateTime.Now
         };
@@ -137,10 +141,11 @@ public class TaskService : ITaskService
         var convertedTaskDto = new TaskDto
         {
             TaskId = taskModel.TaskId,
-            CreationDate = taskModel.CreationDate,
+            StartDate = taskModel.StartDate,
             Description = taskModel.Description,
             Title = taskModel.Title,
-            Status = taskModel.Status.ToString()
+            Status = taskModel.Status.ToString(),
+            Priority = taskModel.Priority.ToString()
         };
         return convertedTaskDto;
     }
@@ -176,10 +181,24 @@ public class TaskService : ITaskService
     /// Deletes a task by id using the TaskRepository
     /// </summary>
     /// <param name="taskId"></param>
-    public void DeleteTask(Guid taskId)
+    public async Task DeleteTask(Guid taskId, bool isOnline )
     {
-        _taskRepository.DeleteTask(taskId);
-    } 
+        if (isOnline)
+        {
+            await _taskRepository.DeleteAsync(taskId);
+            _taskRepository.DeleteTask(taskId);
+        } 
+    }
+ 
+    /// <summary>
+    /// Update the task using TaskRepository 
+    /// </summary>
+    /// <param name="taskDto"></param>
+    public void UpdateTask(TaskDto taskDto)
+    {
+        _taskRepository.UpdateTask(ConvertTaskDtoToTaskModel(taskDto));
+        _taskRepository.UpdateTaskAsync(ConvertTaskDtoToTaskModel(taskDto));
+    }
     
     /// <summary>
     /// Sorts the tasks in a List alphabetically by their title
